Made git not lock optional indices. Closes #12
This allows for users to use git commands, even
if the git segment has not loaded, as it prevents
the git status command from locking the index. This
helps for larger repositories.

Since the git command line flags needed were added
in a relatively recent version of git (2.14.0),
the prompt also automatically detects if the git
version is new enough to support this feature.