Fixed issues with zle-line-init
There was an issue where the color of the arrow, if it was in vi
mode when a command was run, would remain the color of the command
mode arrow, even though it was in insert mode. This was because
the workaround applied for plugin compatibility in #29 actually
just disabled line updating, instead of fixing the actual problem,
which was that the zle widget definitions of old plugins were being
overwritten by our plugin definitions. I fixed this by having
the script generate a wrapper for any preexisting definitions.

As a consequence, it should be noted that any issues that users
might have with widget definitions are probably due to load order.