Fixed some issues with variables not dynamically updating
When configuration variables are changed, the prompt should change,
there were a few settings that did not follow this pattern, so I rewrote
them partially. In addition, for the custom variable prompt I changed
the color so it doesn't visually merge with the neighboring segment.

Also some code style fixes and such.